Removed all migrations to fix bug and re added IntialCreate. To change AliveTime from time to int in db. (time limited to 24hrs)

+++ KollektivSystem.ApiService/Models/Dtos/TicketTypes/CreateTicketTypeRequest.cs
@@ -3,7 +3,7 @@
     public sealed class CreateTicketTypeRequest
     {
         public required string Name { get; init; }
-        public required int Price { get; init; }
+        public required float Price { get; init; }
         public required int AliveTimeMinutes { get; init; }
     }
 }
