+++ KollektivSystem.ApiService/Models/TicketType.cs
@@ -6,7 +6,7 @@
 
         public required string Name { get; init; }
 
-        public required int Price { get; init; }
+        public required float Price { get; init; }
 
         public required TimeSpan AliveTime { get; init; }
         public bool IsActive { get; set; } = true;
